https://github.com/OpenPaaS-Suite/esn-frontend-inbox/issues/53: Moved inboxJmapHelper to esn.inbox.libs
