+++ src/linagora.esn.unifiedinbox/app/services.js
@@ -3,7 +3,7 @@ const _ = require('lodash');
 require('../../esn.inbox.libs/app/services/with-jmap-client/with-jmap-client.js');
 require('../../esn.inbox.libs/app/services/config/config.js');
 require('../../esn.inbox.libs/app/services/jmap-client-wrapper/jmap-client-wrapper.service.js');
-require('./services/jmap-helper/jmap-helper.js');
+require('../../esn.inbox.libs/app/services/jmap-helper/jmap-helper.js');
 require('./services/mailboxes/mailboxes-service.js');
 require('../../esn.inbox.libs/app/services/hook/email-sending-hook.service.js');
 require('../../esn.inbox.libs/app/services/email-body/email-body.js');
