+++ src/linagora.esn.unifiedinbox/app/services/draft/draft.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 require('../../../../esn.inbox.libs/app/services/email-body/email-body.js');
 require('../../services.js');
-require('../../services/jmap-helper/jmap-helper.js');
+require('../../../../esn.inbox.libs/app/services/jmap-helper/jmap-helper.js');
 require('../../../../esn.inbox.libs/app/services/config/config.js');
 
 (function(angular) {
