+++ src/linagora.esn.unifiedinbox/app/services/new-composer/new-composer.js
@@ -1,4 +1,4 @@
-require('../jmap-helper/jmap-helper.js');
+require('../../../../esn.inbox.libs/app/services/jmap-helper/jmap-helper.js');
 
 (function(angular) {
   'use strict';
